REFACTORING :: 파일 위치 변경 및 RacingGame 테스트 이름 모호한 표현 수정
