+++ src/domain/RacingGame/_consts.ts
